opal/util: Move cuda copy code to util from common
As the cuda copy code needs to be included into the
datatype converter, it introduced an open-pal dependency
on the common component. To avoid this, we move the
code into util. This also addresses a bug with
--enable-dynamic-dso and --with-cuda.
