+++ ompi/mca/mtl/base/mtl_base_datatype.h
@@ -28,6 +28,7 @@
 #if OPAL_CUDA_SUPPORT
 #include "opal/mca/common/cuda/common_cuda.h"
 #include "opal/datatype/opal_convertor.h"
+#include "opal/util/opal_cuda_copy.h"
 #endif
 
 #ifndef MTL_BASE_DATATYPE_H_INCLUDED
