+++ ompi/mca/pml/ob1/pml_ob1_recvfrag.c
@@ -55,6 +55,7 @@
 #include "pml_ob1_hdr.h"
 #if OPAL_CUDA_SUPPORT
 #include "opal/mca/common/cuda/common_cuda.h"
+#include "opal/util/opal_cuda_copy.h"
 #endif /* OPAL_CUDA_SUPPORT */
 
 OBJ_CLASS_INSTANCE( mca_pml_ob1_buffer_t,
